Initial extract a FastImageDecoderDecoder
