+++ coil-core/src/androidMain/kotlin/coil3/drawable/ScaleDrawable.kt
