+++ coil-core/src/androidMain/kotlin/coil3/request/gifs.kt
@@ -5,10 +5,10 @@ import android.graphics.drawable.AnimatedImageDrawable
 import android.graphics.drawable.Drawable
 import coil3.Extras
 import coil3.annotation.ExperimentalCoilApi
-import coil3.drawable.MovieDrawable
-import coil3.drawable.MovieDrawable.Companion.REPEAT_INFINITE
 import coil3.getExtra
 import coil3.transform.AnimatedTransformation
+
+private const val REPEAT_INFINITE = -1
 
 /**
  * Set the number of times to repeat the animation if the result is an animated [Drawable].
