+++ coil-core/src/androidMain/kotlin/coil3/transform/AnimatedTransformation.kt
