+++ coil-gif/src/main/java/coil3/util/gifUtils.kt
@@ -1,40 +1,9 @@
 package coil3.util
 
 import android.graphics.Bitmap
-import android.graphics.PixelFormat
-import android.graphics.PostProcessor
-import android.graphics.drawable.Animatable2
 import android.graphics.drawable.Drawable
 import android.os.Build.VERSION.SDK_INT
-import androidx.annotation.RequiresApi
 import androidx.vectordrawable.graphics.drawable.Animatable2Compat
-import coil3.size.Dimension
-import coil3.size.Scale
-import coil3.size.Size
-import coil3.size.isOriginal
-import coil3.size.pxOrElse
-import coil3.transform.AnimatedTransformation
-import coil3.transform.PixelOpacity
-
-@RequiresApi(28)
-internal fun AnimatedTransformation.asPostProcessor() =
-    PostProcessor { canvas -> transform(canvas).flag }
-
-internal val PixelOpacity.flag: Int
-    get() = when (this) {
-        PixelOpacity.UNCHANGED -> PixelFormat.UNKNOWN
-        PixelOpacity.TRANSLUCENT -> PixelFormat.TRANSLUCENT
-        PixelOpacity.OPAQUE -> PixelFormat.OPAQUE
-    }
-
-@RequiresApi(23)
-internal fun animatable2CallbackOf(
-    onStart: (() -> Unit)?,
-    onEnd: (() -> Unit)?
-) = object : Animatable2.AnimationCallback() {
-    override fun onAnimationStart(drawable: Drawable?) { onStart?.invoke() }
-    override fun onAnimationEnd(drawable: Drawable?) { onEnd?.invoke() }
-}
 
 internal fun animatable2CompatCallbackOf(
     onStart: (() -> Unit)?,
@@ -52,18 +21,3 @@ internal inline fun <T> List<T>.forEachIndices(action: (T) -> Unit) {
 
 internal val Bitmap.Config.isHardware: Boolean
     get() = SDK_INT >= 26 && this == Bitmap.Config.HARDWARE
-
-internal inline fun Size.widthPx(scale: Scale, original: () -> Int): Int {
-    return if (isOriginal) original() else width.toPx(scale)
-}
-
-internal inline fun Size.heightPx(scale: Scale, original: () -> Int): Int {
-    return if (isOriginal) original() else height.toPx(scale)
-}
-
-internal fun Dimension.toPx(scale: Scale) = pxOrElse {
-    when (scale) {
-        Scale.FILL -> Int.MIN_VALUE
-        Scale.FIT -> Int.MAX_VALUE
-    }
-}
